Move asset model to eliona package
It is in fact a custom structure for the Eliona package, so it should be
part of it.
